refactor(sidebar/telemetry): move into telemetry module
